refactor: extract to function
